backend/requirements.txt
annotated-types==0.7.0
asgiref==3.7.2
autoflake==2.3.1
autoimport==1.6.0
click==8.1.7
colorama==0.4.6
Django==5.0.2
django-cors-headers==4.3.1
django-filter==23.5
django-simple-history==3.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
maison==1.4.3
Markdown==3.5.2
psycopg==3.1.18
psycopg2==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
pyflakes==3.2.0
PyJWT==2.8.0
pyprojroot==0.3.0
pytz==2024.1
sh==2.0.7
sqlparse==0.4.4
toml==0.10.2
typing_extensions==4.9.0
tzdata==2024.1
xdg==6.0.0
